updated checkForUpdates with error handling, and if statements

+++ main.app/Contents/Resources/Scripts/main.scpt
@@ -30,7 +30,6 @@ on getAppInfo() -- gets the name of the .app file by repeat checking containers 
 		tell application "System Events" to tell application "Finder" to get container of tmpPath as text as alias
 		set tmpPath to result
 		set lastFour to (text ((length of (tmpPath as string)) - 4) through ((length of (tmpPath as string)) - 1) of (tmpPath as string))
-		log tmpPath
 
 		if lastPath is equal to tmpPath then
 			display dialog "Error: couldn't find .app file" with title "Error" buttons ("Okay") default button "Okay"
@@ -51,39 +50,79 @@ end getAppInfo
 
 --------------------------------------------------------------------------------
 on checkForUpdates() -- gets version number from plist file on github and compares with local version number
+	set forward to yes
+
 	try
 		set gitRemote to (do shell script "cd " & POSIX path of appPath & ";git config --get remote.origin.url")
 	on error
 		set gitRemote to result
 		display dialog "Error checking for for updates:
 
-		unable to run command \"git remote -v\" in " & POSIX path of folderPath & "
+unable to run command \"git remote -v\" in " & POSIX path of folderPath & "
 
-		Ensure that the folder containing " & appName & " is linked with github." with title appName buttons ("Okay") default button "Okay"
+Ensure that the folder containing " & appName & " is linked with github." with title appName buttons ("Okay") default button "Okay"
+		set forward to no
+		return "Skip"
 	end try
 
-	set tmpFiles to POSIX path of ((path to temporary items) as text)
-	set currentVersion to do shell script ("osascript -e 'version of app \"" & appPath & "\"'") -- TODO try to get current verstion and display error if can't
-	set plistAddress to "https://raw.githubusercontent.com" & (text (length of "https://github.com/") through (length of gitRemote) of gitRemote) & "/master/" & appName & "/Contents/Info.plist"
+	if forward is yes then
+		set tmpFiles to POSIX path of ((path to temporary items) as text)
+		try
+			set currentVersion to do shell script ("osascript -e 'version of app \"" & appPath & "\"'")
+		on error
+			display dialog "Error checking for updates:
 
-	do shell script "cd " & tmpFiles & "; curl -O " & plistAddress
+unable to read current version" with title appName buttons ("Okay") default button "Okay"
+			set forward to no
+			return "Skip"
+		end try
+	end if
 
-	set newestVersion to do shell script ("defaults read " & tmpFiles & "Info.plist CFBundleShortVersionString")
+	if forward is yes then
+		set plistAddress to "https://raw.githubusercontent.com" & (text (length of "https://github.com/") through (length of gitRemote) of gitRemote) & "/master/" & appName & "/Contents/Info.plist"
+		try
+			do shell script "cd " & tmpFiles & "; curl -O " & plistAddress
+		on error
+			display dialog "Error checking for updates:
 
-	delay 30
+unable to download plist file from " & plistAddress with title appName buttons ("Okay") default button "Okay"
+		set forward to no
+		return "Skip"
+		end try
+	end if
 
-	tell application "System Events"
-		if file (tmpFiles & "/Info.plist") exists then
-			do shell script "rm " & tmpFiles & "/Info.plist"
+	if forward is yes then
+		try
+			set newestVersion to do shell script ("defaults read " & tmpFiles & "Info.plist CFBundleShortVersionString")
+		on error
+			display dialog "Error checking for updates:
+
+unable to read newest version from downloaded plist file" & "
+
+downloaded file from: " & plistAddress with title appName buttons ("Okay") default button "Okay"
+		set forward to no
+		return "Skip"
+		end try
+	end if
+
+	if forward is yes then
+		try
+			tell application "System Events"
+				if file (tmpFiles & "/Info.plist") exists then
+					do shell script "rm " & tmpFiles & "/Info.plist"
+				end if
+			end tell
+		on error
+			display dialog "Error removing old plist file" with title appName buttons ("Okay") default button "Okay"
+		end try
+	end if
+
+	if forward is yes then
+		if currentVersion is not equal to newestVersion then
+			return "Update Available"
+		else if currentVersion is equal to newestVersion then
+			return "No Update Available"
 		end if
-	end tell
-
-	if currentVersion is not equal to newestVersion then
-		log "Update Available"
-		return "Update Available"
-	else if currentVersion is equal to newestVersion then
-		log "No Update Available"
-		return "No Update Available"
 	end if
 end checkForUpdates
 
@@ -106,7 +145,6 @@ end promptForUpdates
 --------------------------------------------------------------------------------
 getAppInfo()
 set appCheck to result
-log appCheck
 if appCheck is not "Skip" then
 	checkForUpdates()
 	set updateCheck to result
@@ -115,7 +153,7 @@ if appCheck is not "Skip" then
 		set userInput to result
 		if userInput is "Update" then
 			try
-				--do shell script "sleep 5; open" & appPath
+			 	-- TODO do shell script "sleep 5; open" & appPath -- for testing (probably need to tell system events to do shell script)
 				quit
 				--do shell script "cd " & folderPath & ";git fetch; sleep 5;git pull;open " &
 			on error
